+++ src/main/kotlin/com/epam/brn/upload/csv/converter/impl/SeriesCsvConverter.kt
@@ -3,7 +3,7 @@ package com.epam.brn.upload.csv.converter.impl
 import com.epam.brn.model.Series
 import com.epam.brn.service.ExerciseGroupsService
 import com.epam.brn.upload.csv.converter.Converter
-import com.epam.brn.upload.csv.dto.SeriesCsv
+import com.epam.brn.upload.csv.record.SeriesCsv
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
