+++ src/main/kotlin/com/epam/brn/upload/csv/iterator/impl/GroupMappingIteratorProvider.kt
@@ -1,7 +1,7 @@
 package com.epam.brn.upload.csv.iterator.impl
 
-import com.epam.brn.upload.csv.dto.GroupCsv
 import com.epam.brn.upload.csv.iterator.MappingIteratorProvider
+import com.epam.brn.upload.csv.record.GroupCsv
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
