+++ src/main/kotlin/com/epam/brn/upload/csv/iterator/impl/SeriesMappingIteratorProvider.kt
@@ -1,7 +1,7 @@
 package com.epam.brn.upload.csv.iterator.impl
 
-import com.epam.brn.upload.csv.dto.SeriesCsv
 import com.epam.brn.upload.csv.iterator.MappingIteratorProvider
+import com.epam.brn.upload.csv.record.SeriesCsv
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
