+++ src/main/kotlin/com/epam/brn/upload/csv/record/GroupCsv.kt
@@ -1,15 +1,12 @@
-package com.epam.brn.upload.csv.dto
+package com.epam.brn.upload.csv.record
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonPropertyOrder
 
-@JsonPropertyOrder(value = ["groupId", "seriesId", "name", "description"])
-data class SeriesCsv(
+@JsonPropertyOrder(value = ["groupId", "name", "description"])
+data class GroupCsv(
     @JsonProperty("groupId")
     val groupId: Long,
-
-    @JsonProperty("seriesId")
-    val seriesId: Long,
 
     @JsonProperty("name")
     val name: String,
